requirements.txt
certifi==2022.12.7
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
elasticsearch==7.15.1
Flask==2.1.0
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
flask-swagger==0.2.14
flask-swagger-ui==4.11.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
PyJWT==2.6.0
pymongo==4.3.3
PyYAML==6.0
requests==2.26.0
six==1.16.0
Unidecode==1.3.6
urllib3==1.26.15
Werkzeug==2.2.3
xmltodict==0.12.0
